refactor: replyCount는 long이므로 그냥 반환하도록 수정

+++ backend/src/main/java/com/ssafy/forest/service/ArticleCommentService.java
@@ -53,7 +53,7 @@ public class ArticleCommentService {
         return articles.map(art -> {
             ArticleComment comment = (ArticleComment) art[0];
             Long replyCount = (Long) art[1];
-            return ArticleCommentResDto.of(comment, articleId, replyCount.intValue());
+            return ArticleCommentResDto.of(comment, articleId, replyCount);
         });
     }
